[FTP-25] Als sporter wil ik personal trainers kunnen zoeken zodat ik een geschikte trainer kan vinden. (#43)
* Created the endpoints for trainer request and also added the app frontend and changed the flow

* Apply suggestions from code review


* Renamed contract

* Removed console logs

* Added some translations and fixed some layout issues

* Fixed the use profile hook

---------

+++ apps/app/src/features/notifications/components/notification-item.tsx
@@ -10,7 +10,7 @@ export const NotificationItem = ({ item }: { item: NotificationItemType }) => {
 	const markAsReadMutator = useMarkNotificationAsRead();
 	const formatter = useFormatter();
 	const isRead = Boolean(item.readAt);
-    
+
 	return (
 		<Pressable
 			key={item.id}
@@ -36,7 +36,7 @@ export const NotificationItem = ({ item }: { item: NotificationItemType }) => {
 					<Text className="text-muted-foreground text-xs">
 						{formatter.dateTime(new Date(item.createdAt), {
 							year: "numeric",
-							month: "short",
+							month: "long",
 							day: "numeric",
 							hour: "2-digit",
 							minute: "2-digit",
